move some struct defination to model mod.
