+++ zihuan_service/src/agent/qq_chat/inbox.rs
@@ -15,7 +15,7 @@ use zihuan_graph_engine::data_value::RedisConfig;
 
 use ims_bot_adapter::adapter::SharedBotAdapter;
 
-use super::core::QqChatAgentService;
+use super::model::QqChatAgentService;
 
 const DEFAULT_CONSUMER_COUNT: usize = 8;
 const REDIS_DEQUEUE_TIMEOUT_SECS: usize = 1;
