+++ zihuan_service/src/agent/qq_chat/model/mod.rs
@@ -1,0 +1,14 @@
+pub(crate) mod context;
+pub(crate) mod inference;
+pub(crate) mod inner;
+pub(crate) mod notifier;
+pub(crate) mod reply;
+
+pub(crate) use context::{QqChatAgentServiceContext, QqChatAgentServiceRuntimeConfig};
+pub(crate) use inference::{QqInferenceToolProvider, QqLoadedInferenceResources};
+pub(crate) use inner::{QqChatAgentService, QqChatAgentServiceInner};
+pub(crate) use notifier::{QqCommandSideEffectContext, QqLongTaskNotifier};
+pub(crate) use reply::{
+    QqChatServiceHandleReport, QqChatServiceReplyBatchBuilder, QqChatServiceReplyBuildRequest,
+    QqChatServiceReplyBuildResult, QqChatServiceTurnResult,
+};
