+++ zihuan_service/src/agent/qq_chat/style_learner.rs
@@ -226,7 +226,7 @@ pub(crate) struct OwnedStyleLearningTaskContext {
     pub intent_classification_llm: Arc<dyn zihuan_core::llm::llm_base::LLMBase>,
     pub rdb_pool: RelationalDbConnection,
     pub max_message_length: usize,
-    pub reply_batch_builder: Option<crate::agent::qq_chat::core::QqChatServiceReplyBatchBuilder>,
+    pub reply_batch_builder: Option<crate::agent::qq_chat::model::QqChatServiceReplyBatchBuilder>,
     pub resolved_language_style_prompt: Option<String>,
 }
