+++ zihuan_service/src/agent/tools/research.rs
@@ -16,7 +16,6 @@ use zihuan_graph_engine::object_storage::S3Ref;
 
 use super::agent_memory::AgentMemoryToolResources;
 use super::common::{optional_string_argument, StaticFunctionToolSpec, ToolNotificationTarget};
-use super::current_time::CurrentTimeBrainTool;
 use super::deep_research::RunDeepResearchSubagentBrainTool;
 use crate::agent::qq_chat::tool_quota::{wrap_brain_tool_with_quota, QqChatToolQuotaContext};
 
@@ -140,7 +139,6 @@ impl BrainTool for RunResearchSubagentBrainTool {
             // The subagent evaluates the problem and either answers directly
             // or escalates to deep_research for multi-step web research.
             let mut brain = Brain::new(Arc::clone(&self.llm));
-            brain.add_tool(CurrentTimeBrainTool);
             brain.add_tool(wrap_brain_tool_with_quota(
                 RunDeepResearchSubagentBrainTool::new(
                     Arc::clone(&self.llm),
